output_connectors/azure_blob_storage/requirements.txt
azure-storage-blob==12.3.0
ubiops-connector==1.0.2
